fix: textBox placeholder 추가
